# chore: optimize avatar names