--- a/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarName/AvatarName.prefab.meta
+++ b/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarName/AvatarName.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 18ee7e67182054d59b2fe9c713890484
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: