--- a/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarShape.asmdef
+++ b/unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Avatar/AvatarShape.asmdef
@@ -38,7 +38,8 @@
         "GUID:c44f6fd10d3d94432a107d581e0096b5",
         "GUID:46cbc133d57a24117993396eb0fbbc5a",
         "GUID:750a3dd4b18d34b46a798e22b150fa10",
-        "GUID:fbcc413e192ef9048811d47ab0aca0c0"
+        "GUID:fbcc413e192ef9048811d47ab0aca0c0",
+        "GUID:c34e38f41494f834abff029ddf82af43"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],