--- a/unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarNamesHUD/AvatarNamesHUDView.cs.meta
+++ b/unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarNamesHUD/AvatarNamesHUDView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e4cb8ad85ada4adc9b05fe6f75a44732
+timeCreated: 1626883596