--- a/unity-renderer/Assets/Scripts/MainScripts/DCL/Models/Players/PlayerModels.asmdef
+++ b/unity-renderer/Assets/Scripts/MainScripts/DCL/Models/Players/PlayerModels.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "PlayerModels"
+}