Fixes #825 

To optimize the Avatar Names we are going to:
- User the same canvas for each name
- Sort the pieces so we can batch the draws of the containers, the text and the voice chat icon.

Approach:
1) Create a new HUD called `AvatarNames`. 
2) Use the `DataStore` to keep a `PlayerStatus` collection.
3) `Minimap` refactoring to use the new `PlayerStatus` collection.
4) `AvatarNameHUD` implementation.


It must be tested with this Kernel: 
decentraland/explorer#2503

Test Link: https://play.decentraland.zone/branch/chore/avatar-names-optimization/index.html?renderer=urn:decentraland:off-chain:renderer-artifacts:chore/optimize-avatar-names