@AjimenezDCL 2021-07-22
# Measurements

## Standing Still
```
Before Optimization
PerformanceMeter - Data report step 2 - Processed values:
 * PERFORMANCE SCORE (0-100) -> 74,49
 * average FPS -> 29,59333
 * highest FPS -> 33
 * lowest FPS -> 25
 * 50 percentile (median) FPS -> 30
 * 95 percentile FPS -> 33
 * total hiccups (>0,05ms frames) -> 3 (0,33% of frames were hiccups)
 * total hiccups time (seconds) -> 0,2143326
 * total frames -> 900
 * total frames time (seconds) -> 30,00287
```

```
After Optimization
PerformanceMeter - Data report step 2 - Processed values:
 * PERFORMANCE SCORE (0-100) -> 75,92
 * average FPS -> 31,42437
 * highest FPS -> 36
 * lowest FPS -> 26
 * 50 percentile (median) FPS -> 32
 * 95 percentile FPS -> 35
 * total hiccups (>0,05ms frames) -> 5 (0,53% of frames were hiccups)
 * total hiccups time (seconds) -> 0,3589172
 * total frames -> 952
 * total frames time (seconds) -> 30,00278
```

## Moving around
```
Before Optimization
PerformanceMeter - Data report step 2 - Processed values:
 * PERFORMANCE SCORE (0-100) -> 73,21
 * average FPS -> 28,47691
 * highest FPS -> 33
 * lowest FPS -> 22
 * 50 percentile (median) FPS -> 29
 * 95 percentile FPS -> 32
 * total hiccups (>0,05ms frames) -> 9 (1,04% of frames were hiccups)
 * total hiccups time (seconds) -> 0,5629196
 * total frames -> 866
 * total frames time (seconds) -> 30,0013
```

```
After Optimization
PerformanceMeter - Data report step 2 - Processed values:
 * PERFORMANCE SCORE (0-100) -> 74,98
 * average FPS -> 30,23965
 * highest FPS -> 35
 * lowest FPS -> 25
 * 50 percentile (median) FPS -> 30
 * 95 percentile FPS -> 35
 * total hiccups (>0,05ms frames) -> 4 (0,44% of frames were hiccups)
 * total hiccups time (seconds) -> 0,2732468
 * total frames -> 918
 * total frames time (seconds) -> 30,0129
```