@AjimenezDCL 2021-07-23
# Measurements 

Benchmark in Build

```
Before Optimization
PerformanceMeter - Data report step 2 - Processed values:
 * PERFORMANCE SCORE (0-100) -> 26.64
 * average FPS -> 18.37345
 * highest FPS -> 19
 * lowest FPS -> 11
 * 50 percentile (median) FPS -> 19
 * 95 percentile FPS -> 19
 * total hiccups (>0.05ms frames) -> 437 (77.35% of frames were hiccups)
 * total hiccups time (seconds) -> 23.67953
 * total frames -> 565
 * total frames time (seconds) -> 30.0313
```

```
After Optimization
PerformanceMeter - Data report step 2 - Processed values:
 * PERFORMANCE SCORE (0-100) -> 39.05
 * average FPS -> 19.19595
 * highest FPS -> 20
 * lowest FPS -> 11
 * 50 percentile (median) FPS -> 19
 * 95 percentile FPS -> 20
 * total hiccups (>0.05ms frames) -> 319 (53.89% of frames were hiccups)
 * total hiccups time (seconds) -> 16.94389
 * total frames -> 592
 * total frames time (seconds) -> 30.02672
```