@AjimenezDCL 2021-07-26
> LGTM! Some suggestions were added.
> 
> I think we should somehow ( maybe each X seconds ) change the names of the avatars when we are moving and we are above the max numbers of avatar names.
> 
> Imagine that you are in a concert and we are moving on to where our friends are. So we are moving from the entrance to the other part of the concert. You will want to see the avatar names of the nears players and not the first 200 avatars names.
> 
> Take this into account for the future!

@Suduck Yeah I was thinking the same but as I started benchmarking I noticed we can drastically increase the number of Avatar Names in screen (thanks the the batching of the UI parts). It's a nice suggestion but it wouldn't have much use until we can comfortably have more than 200 users on screen.